Former-commit-id: efa2cd23ec4d05ff5c6047eec8c773749539adfc

+++ Sources/Internal/Network/Private/ProtoDecoder.cpp
@@ -73,7 +73,7 @@ size_t ProtoDecoder::EncodeDataFrame(ProtoHeader* header, uint32 channelId, uint
     header->frameType = TYPE_DATA;
     header->channelId = channelId;
     header->packetId = packetId;
-    header->totalSize = packetSize;
+    header->totalSize = static_cast<uint32>(packetSize);
     return sizeToEncode;
 }
